pyproject.toml
[project]
name = "llm-structured-semantic-generation"
version = "0.1.0"
description = "TFM project"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "plotly>=6.3.1",
    "pyyaml>=6.0.2",
    "jsonschema>=4.25.1",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
notebooks = [
    "ipykernel>=6.30.1",
    "jupyterlab>=4.4.7",
]
llm = [
    "accelerate>=1.10.1",
    "bitsandbytes>=0.43.0",
    "datasets>=4.1.1",
    "evaluate>=0.4.6",
    "peft>=0.17.1",
    "safetensors>=0.6.2",
    "sentencepiece>=0.2.1",
    "torch>=2.5.0",
    "transformers>=4.56.1",
    "trl>=0.23.1",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
torchaudio = { index = "pytorch-cu121" }
